refactor(snowman): extract model construction (R3.8)
