+++ docs/ARCHITECTURE.md
@@ -239,10 +239,11 @@ camera.position -= shake                          // revert so smoothing stays c
 > Stage R3 (in progress) splits `snowman.ts`: the implementation moved to
 > **`src/snowman/index.ts`** behind a thin **`src/snowman.ts`** facade
 > (`export * from './snowman/index.js'`) so every `./snowman.js` importer keeps
-> resolving a sibling file. `index.ts` stays free of *relative* imports so the
-> bare-Node physics-invariant harness can load it directly (it imports
-> `src/snowman/index.ts`, not the facade). Steps 8–12 carve `model` / `pose` /
-> `physics` / `collision` / `test-hooks` out of `index.ts`.
+> resolving a sibling file. R3.8 moved model construction into
+> **`src/snowman/model.ts`**; the verification harness self-registers the same
+> `.js` -> `.ts` resolver as the Node suites before importing the facade, so the
+> public seam remains the thing under test as steps 9–12 carve `pose` / `physics` /
+> `collision` / `test-hooks` out of `index.ts`.
 
 ---
